backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
transformers>=4.44.2
torch>=2.7.1
regex>=2024.5.15
python-dotenv>=1.0.1
